reorganizing topography functions
